Add files via upload
This project is a GUI coded in Java (using Eclipse). It includes a class that defines the Object Ball, which is a filled in circle. A new ball can be created by pressing the + sign in the top left and deleted by pressing the - minus sign (minimum of 1 ball and maximum of 10). Adding a ball will reset the position of all of the balls to their original coordinates.

Running the main file will start the project.

If a ball is clicked on and the mouse is dragged, the ball will follow the cursor until it is released. 
When a ball is released, it will continue to travel with the same speed it had when it was released. I used a timer to do this: for every ball that is visible, there is another ball, which is not drawn, that is in the same position as its corresponding ball was 10 milliseconds prior. The difference in X and Y positions of these balls determines the X and Y speeds.

There is a slider to control acceleration. Y speed and X speed will accelerate by this slider, either in the negative direction (left) or the positive direction (right).

Pressing the reset button will reset all of the balls.

When either an untouched ball or a ball that is currently being held touches a moving ball, the moving ball will bounce off in the opposite direction (very simple collision).